chore: s2s v2 formal chain
