+++ packages/shared/__tests__/components/SubscanLink.spec.tsx
@@ -2,7 +2,7 @@ import { render } from '@testing-library/react';
 import React from 'react';
 import { SubscanLink } from '../../components/widget/SubscanLink';
 import { SYSTEM_ChAIN_CONFIGURATIONS } from '../../config/network';
-import { knownDVMNetworks, knownPolkadotNetworks } from '../../config/network/category';
+import { knownDVMNetworks, knownPolkadotNetworks } from '../../config/network';
 import { DVMNetwork, Network, PolkadotTypeNetwork } from '../../model';
 
 describe('<SubscanLink />', () => {
